// src/types/Synergetic/Finance/iSynVDonorReceipt.ts
type iSynVDonorReceipt = {
  ReceiptSeq: number;
  PledgeSeq: number;
  DonorID: number;
  DonorIDJoint: number;
  DonorName: string;
  DonorNameRaw: string;
  DonorMailName: string | null;
  DonorMailSalutation: string | null;
  DonorNameInternal: string | null;
  DonorNameExternal: string | null;
  DonorLegalFullName: string | null;
  DonorSurname: string | null;
  DonorPreferred: string | null;
  DonorPreferredFormal: string | null;
  DonorGiven1: string | null;
  DonorGiven2: string | null;
  DonorInitials: string | null;
  DonorAddressFull: string | null;
  DonorAddressDPIDBarcode: string | null;
  DonorAddressComma: string;
  DonorHomePhone: string;
  DonorHomeFax: string;
  DonorOccupPhone: string | null;
  DonorOccupFax: string | null;
  DonorMobilePhone: string | null;
  DonorDefaultMobilePhone: string | null;
  DonorOccupMobilePhone: string | null;
  DonorHomeEmail: string | null;
  DonorOccupEmail: string | null;
  DonorDeceasedFlag: boolean | null;
  ReceiptNo: number;
  ReceiptAmount: number;
  ReceiptPaymentMethodCode: string;
  ReceiptDrawer: string | null;
  ReceiptDate: Date | string | null;
  ReceiptAppeal: string;
  ReceiptAppealDescription: string | null;
  ReceiptFund: string | null;
  ReceiptFundDescription: string | null;
  ReceiptFundTaxDeductableFlag: boolean | null;
  ReceiptAnonymousFlag: boolean | null;
  ReceiptRecognitionFlag: boolean | null;
  ReceiptComments: string;
  PostingNumber: number;
  ReceiptNumber: number;
  DonorSpouseSurname: string | null;
  DonorSpousePreferred: string | null;
  DonorSpouseGiven1: string | null;
  DonorSpouseGiven2: string | null;
  DonorSpouseID: number | null;
  DonorDefaultEmail: string | null;
  DonorOccupDesc: string | null;
  DonorSpouseDefaultEmail: string | null;
  DonorSpouseOccupDesc: string | null;
}

export default iSynVDonorReceipt;